Rename files for storybook
